Added octree for modified chunks, added saving and loading

+++ src/engine/savegame.c
@@ -36,12 +36,12 @@ uint8_t openSave(char* folder, char* name, uint8_t writing)
     }
 }
 
-void writeElement(void* ptr, uint16_t size)
+void writeElement(void* ptr, size_t size)
 {
     fwrite(ptr, size, 1, saveFile);
 }
 
-void readElement(void* ptr, uint16_t size)
+void readElement(void* ptr, size_t size)
 {
     fread(ptr, size, 1, saveFile);
 }
