+++ src/engine/savegame.h
@@ -1,11 +1,12 @@
 #ifndef _SAVEGAME_H
 #define _SAVEGAME_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 uint8_t openSave(char* folder, char* name, uint8_t writing);
-void writeElement(void* ptr, uint16_t size);
-void readElement(void* ptr, uint16_t size);
+void writeElement(void* ptr, size_t size);
+void readElement(void* ptr, size_t size);
 void closeSave();
 
 #endif
